Fixed eslint.config.js path and override
 - Also removed unused tsconfig properties
