+++ eslint.config.js
@@ -25,8 +25,8 @@ module.exports = [
       ecmaVersion: 2023,
       sourceType: 'commonjs',
       parserOptions: {
-        // https://typescript-eslint.io/blog/parser-options-project-true/
-        project: true
+        project: true,
+        tsconfigRootDir: __dirname
       }
     },
     rules: {
