+++ templates/node/eslint.config.js
@@ -1,1 +1,10 @@
-module.exports = [...require('./node_modules/@connectedcars/setup/eslint.config.js')]
+module.exports = [
+  ...require('./node_modules/@connectedcars/setup/eslint.config.js'),
+  {
+    languageOptions: {
+      parserOptions: {
+        tsconfigRootDir: __dirname
+      }
+    }
+  }
+]
